API crash when 'format' JSON schema is triggered (#5795)
In this case, a user was created with an empty ("") email address. It
failed both the minLength and format requirements but kz_json_schema
was not ready to handle the wrong_format error from jesse and the API
request crashed.

This PR adds wrong_format handling as well as introduces user API CRUD
testing in kazoo_proper. Also increases minLength for emails to 3 (1
makes no sense).

Additionally, crashes occurred during user validation if schema
validation failed. The functions that tried to modify the user
doc (loaded after successful validation) failed because failed
validation did not set a doc in cb_context and thus kz_json operations
occurred against 'undefined'. This PR sets those apart and checks that
validation of the request was successful before running the updates.

Introduced a complement to kz_json:exec/2: kz_json:exec_first/2 which
runs the functions with the JObj accumulator in the first position of
the supplied functions (versus in the last position as in exec/2.

Also updates all pqc API modules to use full expectations instead of
the shorthand; made dialyzer much happier.