HELP-44667: don't send authz resp on quick hangup (#5914)
* HELP-44667: don't send authz resp on quick hangup

Processing the authz request can take longer than processing the
subsequent CHANNEL_DESTROY. This causes the flushing in j5_channels to
occur before the storing of the authorization payload. Since the
channel is gone, the authorization payload isn't flushed and counts
against the account/reseller limits.

Instead, check the j5_channels ETS table for the authorization
payload on CHANNEL_DESTROY; if none, store an entry with a
`destroyed`:`true` flag. If a subsequent authz response is attempted
to be stored, this check will keep it from happening.

Introduce a test for sending CHANNEL_DESTROY followed by authz_req and
ensuring totals remain 0 and no authz_resp is generated.

* edocified

* update app file minus kazoo applications

* cheat the system

* enough uncircling

* set timestamp on channel_destroy test

* move to public table and ets ops outside gen_server when able

* specs and formatting

* specs for all